+++ wild-bind-indicator/ChangeLog.md
@@ -3,6 +3,7 @@
 ## 1.0.0.0  -- ?
 
 * Now it uses `gi-gtk` package, instead of `gtk` package. `gi-gtk` package has different prerequisites than `gtk`.
+* Now it uses `async` package to clean-up the WildBind action when the Gtk+ GUI finishes. Probably fixes #4 and #5a.
 
 ## 0.2.0.0  -- 2018-01-01